% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{ANOVA}
\alias{ANOVA}
\alias{print.ANOVA}
\title{One-way and Two-way ANOVA}
\usage{
ANOVA(data, variable, group, conf.level = 0.95)
}
\arguments{
\item{data}{a data frame containing the variables.}

\item{variable}{a character string specifying the numeric variable of
interest.}

\item{group}{a character vector specifying one or two grouping variables.}

\item{conf.level}{a number between 0 and 1 giving the confidence level of
the confidence interval.}
}
\value{
An object of class \code{ANOVA}.  The \code{print} method produces
a LaTeX table that mimics the look of SPSS output (version <24).
}
\description{
Perform one-way or two-way ANOVA on variables of a data set.  The output is
printed as a LaTeX table that mimics the look of SPSS output (version <24).
}
\author{
Andreas Alfons
}
\keyword{htest}
