% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqTest.R
\name{chisqTest}
\alias{chisqTest}
\alias{print.chisqTest}
\title{\eqn{\chi^{2}}{Chi-squared} Tests}
\usage{
chisqTest(data, variables, p = NULL)
}
\arguments{
\item{data}{a data frame containing the variables.}

\item{variables}{a character vector specifying the categorical variable(s)
of interest.  If only one variable is specified, a goodness-of-fit test is
performed.  If two variables are specified, a test on independence is
performed (with the first variable used for the rows and the second variable
for the columns of the crosstabulation).}

\item{p}{a vector of probabilities for the categories in the
goodness-of-fit test.}
}
\value{
An object of class \code{chisqTest}.  The \code{print} method
produces a LaTeX table that mimics the look of SPSS output (version <24).
}
\description{
Perform a \eqn{\chi^{2}}{chi-squared} goodness-of-fit test or a
\eqn{\chi^{2}}{chi-squared} test on independence on variables of
a data set.  The output is printed as a LaTeX table that mimics
the look of SPSS output (version <24).
}
\note{
The test on independence also reports the results of a likelihood
ratio test.
}
\author{
Andreas Alfons
}
\keyword{htest}
