% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2spss-package.R
\docType{package}
\name{r2spss-package}
\alias{r2spss-package}
\alias{r2spss}
\title{\packageTitle{r2spss}}
\description{
\packageDescription{r2spss}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{r2spss}
\packageIndices{r2spss}
}
\author{
\packageAuthor{r2spss}

Maintainer: \packageMaintainer{r2spss}
}
\keyword{package}
