% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2spss.sty.R
\name{r2spss.sty}
\alias{r2spss.sty}
\title{Create the LaTeX style file for 'r2spss'}
\usage{
r2spss.sty(path = NULL)
}
\arguments{
\item{path}{a character string specifying the path to the folder in which
to put the style file, or \code{NULL} (the default) to print the contents
of the style file to the standard output connection (usually the \R
console).}
}
\value{
Nothing is returned, the function is called for its side effects.
}
\description{
Create the LaTeX style file required to compile LaTeX documents that include
tables created by package \pkg{r2spss}.  You can put the resulting file
\emph{r2spss.sty} in the folder containing your LaTeX document, and you
should include \preformatted{\usepackage{r2spss}} in the preamble of your
LaTeX document.
}
\examples{
# print contents of style file
r2spss.sty()

\dontrun{

  # put file 'r2spss.sty' in the current working directory
  r2spss.sty(".")
}

}
\author{
Andreas Alfons
}
\keyword{IO}
