% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSPSS.R
\name{toSPSS}
\alias{toSPSS}
\title{Convert R objects to SPSS-style tables}
\usage{
toSPSS(object, ...)
}
\arguments{
\item{object}{an \R object for which a \code{toSPSS} method exists, such
as objects returned by functions in \pkg{r2spss}.}

\item{\dots}{additional arguments passed down to methods.}
}
\value{
In order to work as expected, methods of \code{toSPSS} should return an
object of class \code{"SPSSTable"}.  It should include a component
\code{table} that contains a data frame, which can be supplied as the
first argument to \code{\link[=toLatex.data.frame]{toLatex}} to print
a LaTeX table that mimics the look of SPSS output.  Additional components
of the returned object define additional arguments to be passed to the
\code{"data.frame"} method of \code{\link[=toLatex.data.frame]{toLatex}}.
}
\description{
Generic function to convert an \R object into an object that contains all
necessary information for printing a LaTeX table that mimics the look of
SPSS output.
}
\examples{
# load data
data("Eredivisie")

# compute a Kruskual-Wallis test to investigate whether
# market values differ by playing position
kw <- kruskalTest(Eredivisie, "MarketValue",
                  group = "Position")

# convert to an object of class "SPSSTable" that
# contains the table with the test results
kwSPSS <- toSPSS(kw, statistics = "test")
kwSPSS

# blank out the number of degrees of freedom to ask
# an assignment question about it
kwSPSS$table[2, 1] <- "???"

# print the LaTeX table to be included in the assignment
toLatex(kwSPSS)

}
\author{
Andreas Alfons
}
\keyword{manip}
