% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTest.R
\name{tTest}
\alias{tTest}
\alias{print.tTest}
\title{t Tests}
\usage{
tTest(data, variables, group = NULL, mu = 0, conf.level = 0.95)

\method{print}{tTest}(x, digits = 3, statistics = c("statistics",
  "test"), ...)
}
\arguments{
\item{data}{a data frame containing the variables.}

\item{variables}{a character vector specifying numeric variable(s) to be
used for testing the mean(s).  If \code{group} is \code{NULL}, a one-sample
t test is performed if only one variable is specified, and a paired-samples
t test is performed if two variables are specified.  If a grouping variable
is specified in \code{group}, an independent-samples t-test is performed and
this should be a character string specifying the numeric variable of
interest.}

\item{group}{a character string specifying a grouping variable for an
independent-samples t-test, or \code{NULL}.}

\item{mu}{a number indicating the true value of the mean for a one-sample
t test.}

\item{conf.level}{a number between 0 and 1 giving the confidence level of
the confidence interval.}

\item{x}{an object of class \code{"tTest"} as returned by function
\code{tTest}.}

\item{digits}{an integer giving the number of digits after the comma to be
printed in the LaTeX tables.}

\item{statistics}{a character vector specifying which LaTeX tables should
be printed.  Available options are \code{"statistics"} for descriptive
statistics and \code{"test"} for test results.  The default is to print both
tables.}

\item{\dots}{currently ignored.}
}
\value{
An object of class \code{tTest}.  The \code{print} method produces
a LaTeX table that mimics the look of SPSS output (version <24).
}
\description{
Perform an one-sample t test, a paired-samples t test or an
independent-samples t test on variables of a data set.  The output
is printed as a LaTeX table that mimics the look of SPSS output
(version <24).
}
\author{
Andreas Alfons
}
\keyword{htest}
