% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxonTest.R
\name{wilcoxonTest}
\alias{wilcoxonTest}
\alias{print.wilcoxonTest}
\title{Wilcoxon Signed Rank and Rank Sum Tests}
\usage{
wilcoxonTest(data, variables, group = NULL)

\method{print}{wilcoxonTest}(x, digits = 2:3, statistics = c("ranks",
  "test"), ...)
}
\arguments{
\item{data}{a data frame containing the variables.}

\item{variables}{a character vector specifying numeric variable(s) to be
used.  If \code{group} is \code{NULL}, the Wilcoxon signed rank test is
performed and this should be a character vector specifying two numeric
variables which contain the paired observations.  If a grouping variable is
specified in \code{group}, the Wilcoxon rank sum test is performed and this
should be a character string specifying the numeric variable of interest.}

\item{group}{a character string specifying a grouping variable for the
Wilcoxon rank sum test, or \code{NULL}.}

\item{x}{an object of class \code{"wilcoxonTest"} as returned by function
\code{wilcoxonTest}.}

\item{digits}{an integer vector giving the number of digits after the comma
to be printed in the LaTeX tables.  The first element corresponds to the
number of digits in table with the summary of the ranks, and the second
element corresponds to the number of digits in the table for the test.}

\item{statistics}{a character vector specifying which LaTeX tables should
be printed.  Available options are \code{"ranks"} for a summary of the ranks
and \code{"test"} for test results.  The default is to print both tables.}

\item{\dots}{currently ignored.}
}
\value{
An object of class \code{wilcoxonTest}.  The \code{print} method
produces a LaTeX table that mimics the look of SPSS output (version <24).
}
\description{
Perform a Wilcoxon signed rank test for paired samples or a Wilcoxon rank
sum test for independent samples on variables of a data set.  The output
is printed as a LaTeX table that mimics the look of SPSS output (version
<24).
}
\note{
The Wilcoxon rank sum test also reports the value of the equivalent
Mann-Whitney U test statistic.
}
\author{
Andreas Alfons
}
\keyword{htest}
