% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqTest.R
\name{chisqTest}
\alias{chisqTest}
\alias{print.chisqTest}
\title{\eqn{\chi^{2}}{Chi-squared} Tests}
\usage{
chisqTest(data, variables, p = NULL)

\method{print}{chisqTest}(x, digits = c(1, 3),
  statistics = c("frequencies", "test"), ...)
}
\arguments{
\item{data}{a data frame containing the variables.}

\item{variables}{a character vector specifying the categorical variable(s)
of interest.  If only one variable is specified, a goodness-of-fit test is
performed.  If two variables are specified, a test on independence is
performed (with the first variable used for the rows and the second variable
for the columns of the crosstabulation).}

\item{p}{a vector of probabilities for the categories in the
goodness-of-fit test.}

\item{x}{an object of class \code{"chisqTest"} as returned by function
\code{chisqTest}.}

\item{digits}{an integer vector giving the number of digits after the comma
to be printed in the LaTeX tables.  The first element corresponds to the
number of digits in the table of frequencies, and the second element
corresponds to the number of digits in the table for the test.}

\item{statistics}{a character vector specifying which LaTeX tables should
be printed.  Available options are \code{"frequencies"} for a table of the
observed and expected frequencies, and \code{"test"} for test results.  The
default is to print both tables.}

\item{\dots}{currently ignored.}
}
\value{
An object of class \code{chisqTest}.  The \code{print} method
produces a LaTeX table that mimics the look of SPSS output (version <24).
}
\description{
Perform a \eqn{\chi^{2}}{chi-squared} goodness-of-fit test or a
\eqn{\chi^{2}}{chi-squared} test on independence on variables of
a data set.  The output is printed as a LaTeX table that mimics
the look of SPSS output (version <24).
}
\note{
The test on independence also reports the results of a likelihood
ratio test.
}
\author{
Andreas Alfons
}
\keyword{htest}
