% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\alias{print.tTest}
\title{Descriptive Statistics}
\usage{
descriptives(data, variables)
}
\arguments{
\item{data}{a data frame containing the variables.}

\item{variables}{a character vector specifying numeric variables for which
to compute descriptive statistics.}
}
\value{
An object of class \code{descriptives}.  The \code{print} method
produces a LaTeX table that mimics the look of SPSS output (version <24).
}
\description{
Compute descriptive statistics of numeric variables of a data set (number of
observations, minimum, maximum, mean, standard deviaiton).  The output is
printed as a LaTeX table that mimics the look of SPSS output (version <24).
}
\author{
Andreas Alfons
}
\keyword{univar}
