% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskalTest.R
\name{kruskalTest}
\alias{kruskalTest}
\alias{print.kruskalTest}
\title{Kruskal-Wallis Test}
\usage{
kruskalTest(data, variable, group)
}
\arguments{
\item{data}{a data frame containing the variables.}

\item{variable}{a character string specifying the numeric variable of
interest.}

\item{group}{a character string specifying a grouping variable.}
}
\value{
An object of class \code{kruskalTest}.  The \code{print} method
produces a LaTeX table that mimics the look of SPSS output (version <24).
}
\description{
Perform a Kruskal-Wallis test on variables of a data set.  The output is
printed as a LaTeX table that mimics the look of SPSS output (version <24).
}
\author{
Andreas Alfons
}
\keyword{htest}
