% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{paletteSPSS}
\alias{paletteSPSS}
\title{SPSS Color Palette}
\usage{
paletteSPSS(n = NULL, version = r2spssOptions$get("version"))
}
\value{
A character vector specifying 30 colors as used by SPSS.
}
\description{
Color palette used by SPSS (e.g., for multiple lines in a plot).
}
\examples{
df <- data.frame(x = 1:30, y = 0)
colors <- paletteSPSS()
plotSPSS(df, c("x", "y"), col = colors, pch = 16)

}
\author{
Andreas Alfons
}
\keyword{color}
