% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\alias{print.regression}
\title{Linear Regression}
\usage{
regression(..., data, labels = NULL)
}
\arguments{
\item{\dots}{at least one formula specifying a regression models.  Different
models can be compared by supplying multiple formulas.}

\item{data}{a data frame containing the variables.}

\item{labels}{a character or numeric vector giving labels for the
regression models in the output tables.}
}
\value{
An object of class \code{regression}.  The \code{print} method
produces a LaTeX table that mimics the look of SPSS output (version <24).
}
\description{
Perform linear regression variables of a data set.  The output is printed as
a LaTeX table that mimics the look of SPSS output (version <24).
}
\author{
Andreas Alfons
}
\keyword{multivariate}
