% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signTest.R
\name{signTest}
\alias{signTest}
\alias{toSPSS.signTestSPSS}
\alias{print.signTestSPSS}
\title{Sign Test}
\usage{
signTest(data, variables, exact = FALSE)

\method{toSPSS}{signTestSPSS}(
  object,
  statistics = c("test", "frequencies"),
  version = c("modern", "legacy"),
  ...
)

\method{print}{signTestSPSS}(
  x,
  statistics = c("frequencies", "test"),
  version = c("modern", "legacy"),
  ...
)
}
\arguments{
\item{data}{a data frame containing the variables.}

\item{variables}{a character vector specifying two numeric variables
containing the paired observations.}

\item{exact}{a logical indicating whether or not to include the exact
p-value using the binomial distribution.  Note that the p-value using the
normal approximation is always reported.}

\item{object, x}{an object of class \code{"signTestSPSS"} as returned by
function \code{signTest}.}

\item{statistics}{a character string or vector specifying which SPSS tables
to produce.   Available options are \code{"frequencies"} for a summary of
the frequencies and \code{"test"} for test results.  For the \code{toSPSS}
method, only one option is allowed (the default is the table of test
results), but the \code{print} method allows several options (the default
is to print all tables).}

\item{version}{a character string specifying whether the table should
mimic the content and look of recent SPSS versions (\code{"modern"}) or
older versions (<24; \code{"legacy"}).  The main difference in terms of
content is that small p-values are displayed differently.}

\item{\dots}{additional arguments to be passed down to
\code{\link{formatSPSS}}.}
}
\value{
An object of class \code{"signTestSPSS"} with the following
components:
\describe{
  \item{\code{statistics}}{a data frame containing information on the
  number of observations with negative and positive differences.}
  \item{\code{asymptotic}}{a list containing the results of the test using
  the normal approximation.}
  \item{\code{exact}}{if requested, a numeric vector containing the exact
  two-sided p-value, one-sided p-value, and point probability using the
  binomial distribution.}
  \item{\code{variables}}{a character vector containing the names of the two
  numeric variables with the paired observations.}
  \item{\code{n}}{an integer giving the number of observations.}
}

The \code{toSPSS} method returns an object of class \code{"SPSSTable"}
which contains all relevant information in the required format to produce
the LaTeX table.  See \code{\link[=toLatex.SPSSTable]{toLatex}} for possible
components and how to further customize the LaTeX table based on the
returned object.

The \code{print} method produces a LaTeX table that mimics the look of SPSS
output.
}
\description{
Perform a sign test for a paired sample on variables of a data set.  The
output is printed as a LaTeX table that mimics the look of SPSS output.
}
\details{
The \code{print} method first calls the \code{toSPSS} method followed by
\code{\link[=toLatex.SPSSTable]{toLatex}}.  Further customization can be
done by calling those two functions separately, and modifying the object
returned by \code{toSPSS}.
}
\note{
LaTeX tables that mimic recent versions of SPSS (\code{version = "modern"})
may require several LaTeX compilations to be displayed correctly.
}
\examples{
# load data
data("Exams")

# test whether grades differ between the
# regular exam and the resit
signTest(Exams, c("Regular", "Resit"))

}
\author{
Andreas Alfons
}
\keyword{htest}
