% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxonTest.R
\name{wilcoxonTest}
\alias{wilcoxonTest}
\alias{print.wilcoxonTest}
\title{Wilcoxon Rank Sum and Signed Rank Tests}
\usage{
wilcoxonTest(data, variables, group = NULL)
}
\arguments{
\item{data}{a data frame containing the variables.}

\item{variables}{a character vector specifying numeric variable(s) to be
used.  If \code{group} is \code{NULL}, the Wilcoxon signed rank test is
performed and this should be a character vector specifying two numeric
variables which contain the paired observations.  If a grouping variable is
specified in \code{group}, the Wilcoxon rank sum test is performed and this
should be a character string specifying the numeric variable of interest.}

\item{group}{a character string specifying a grouping variable for the
Wilcoxon rank sum test, or \code{NULL}.}
}
\value{
An object of class \code{wilcoxonTest}.  The \code{print} method
produces a LaTeX table that mimics the look of SPSS output (version <24).
}
\description{
Perform a Wilcoxon signed rank test for paired samples or a Wilcoxon rank
sum test for independent samples on variables of a data set.  The output
is printed as a LaTeX table that mimics the look of SPSS output (version
<24).
}
\note{
The Wilcoxon rank sum test also reports the value of the equivalent
Mann-Whitney U test statistic.
}
\author{
Andreas Alfons
}
\keyword{htest}
